requirements.txt
streamlit
keras
tensorflow
pillow
opencv-python-headless
numpy
